chore: update package metadata and rename package folder
