+++ MANIFEST.in
@@ -1,4 +1,4 @@
-include ./tdd_api_plugin_aas/*.py
+include ./domus_tdd_api_plugin_aid/*.py
 
 include *.md
 include tox.ini
@@ -9,10 +9,10 @@ exclude .gitlab-ci.yml
 exclude .gitlab-ci-extended.yml
 
 exclude pytest.ini
-prune tdd_api_plugin_aas/tests
-recursive-include tdd_api_plugin_aas *.jinja2
-recursive-include tdd_api_plugin_aas *.json
-recursive-include tdd_api_plugin_aas *.n3
-recursive-include tdd_api_plugin_aas *.trig
-recursive-include tdd_api_plugin_aas *.ttl
-recursive-include tdd_api_plugin_aas *.xml
+prune domus_tdd_api_plugin_aid/tests
+recursive-include domus_tdd_api_plugin_aid *.jinja2
+recursive-include domus_tdd_api_plugin_aid *.json
+recursive-include domus_tdd_api_plugin_aid *.n3
+recursive-include domus_tdd_api_plugin_aid *.trig
+recursive-include domus_tdd_api_plugin_aid *.ttl
+recursive-include domus_tdd_api_plugin_aid *.xml
