+++ domus_tdd_api_plugin_aid/tests/data/aas/one_aid.trig
