+++ domus_tdd_api_plugin_aid/tests/data/td-to-aas/td_aas.trig
